% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fun_2_load_data}
\alias{fun_2_load_data}
\title{Download datasets from GitHub}
\usage{
fun_2_load_data(location, filename, url, overwrite = FALSE)
}
\arguments{
\item{location}{the string name of the sub-repository location to load the data}

\item{filename}{the txt filename of the data to load}

\item{url}{the string GitHub url for the data txt file}

\item{overwrite}{a \code{logical}. If \code{TRUE}, the file will be downloaded again
and the previous version will be replaced.}
}
\value{
No return value
}
\description{
This function downloads the datasets hosted on any
GitHub repository . The file won't be downloaded if already exists
locally (except if \code{overwrite = TRUE}).
The file will be stored in the chosen sub-repository in
\verb{data/}. This folder will be created if required.
}
